requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.4
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.2.1
django-rest-multiple-models==2.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
drfdocs==0.0.11
Faker==14.2.0
filelock==3.6.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
pbr==5.8.1
Pillow==9.1.1
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
validate-email==1.3
virtualenv==20.14.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
